--- public/index.html
@@ -1,51 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta
-      name="viewport"
-      content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0"
-    />
-    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    <link rel="stylesheet" href="css/normalize.css" />
-    <link rel="stylesheet" href="css/style.css" />
-    <title>TASKMANAGER</title>
-  </head>
-
-  <body>
-    <div style="display: none">
-      <svg>
-        <symbol
-          id="wave"
-          viewBox="0 0 159 10"
-          preserveAspectRatio="none"
-        >
-          <g
-            fill="none"
-            fill-rule="nonzero"
-            stroke-linecap="square"
-            stroke-width="5"
-          >
-            <path
-              d="M4 4l9.257 2.463L21.367 4l7.927 2.463L38.736 4l9.117 2.463L56.103 4l8.685 2.463L73.472 4l8.684 2.463L90.84 4l8.684 2.463L108.208 4l8.684 2.463L125.576 4l8.684 2.463L142.943 4M146.423 4l9.257 2.463"
-            />
-          </g>
-        </symbol>
-      </svg>
-    </div>
-
-    <main class="main">
-      <section class="main__control control container">
-        <h1 class="control__title">TASKMANAGER</h1>
-        <!-- Меню -->
-      </section>
-
-      <!-- Поиск -->
-
-      <!-- Фильтры -->
-
-      <!-- Контент -->
-    </main>
-    <script src="bundle.js"></script>
-  </body>
-</html>
