+++ src/components/more-button.js
@@ -1,0 +1,3 @@
+export const createMoreButtonTemplate = (title) => {
+  return `<button class="load-more" type="button">${title}</button>`;
+};
